% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{add_exhausted}
\alias{add_exhausted}
\title{Adds correct exhausted numbers for a `rcv_tally` dataframe}
\usage{
add_exhausted(results)
}
\arguments{
\item{results}{A dataframe that comes from `rcv_tally()`}
}
\value{
The results dataframe with correct counts for the exhausted votes
}
\description{
Adds correct exhausted numbers for a `rcv_tally` dataframe
}
