% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox_functions.R
\name{ballot_tidy}
\alias{ballot_tidy}
\title{Function for reading in CA style ballot images}
\usage{
ballot_tidy(rawimage, key, df)
}
\arguments{
\item{rawimage}{File containing the raw ballot image}

\item{key}{File containing the lookup file}

\item{df}{True/False depending on whether the rawimage and key are stored as dataframes}
}
\value{
A dataframe that can be easily read and used for calculating results
}
\description{
Function for reading in CA style ballot images
}
\examples{
ballot_tidy(rawimage = "http://www.acgov.org/rov/rcv/results/230/BerkeleyMayor/ballot_image.txt",
key = "http://www.acgov.org/rov/rcv/results/230/BerkeleyMayor/master_lookup.txt", df = F)
ballot_tidy(rawimage = sf_bos_ballot, key = sf_bos_lookup, df = T)
}
