% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cambridge_ballot}
\alias{cambridge_ballot}
\title{Raw ballot image data from a Cambridge, MA RCV election}
\format{A data frame with 17959 rows and 26 variables}
\usage{
cambridge_ballot
}
\description{
The .rda version of a .csv from a 2005 Cambridge, MA City Council
multiwinner RCV election. This data is included as an example of
another ballot image format for use with read_ballot.
}
\keyword{datasets}
