% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{cambridge_lookup}
\alias{cambridge_lookup}
\title{Raw lookup data from a Cambridge, MA RCV election}
\format{A data frame with 46 rows and 2 variables}
\usage{
cambridge_lookup
}
\description{
The .rda version of a .csv from a 2005 Cambridge, MA City Council
multiwinner RCV election. This data is included as an example of
another master lookup format for use with read_lookup.
}
\keyword{datasets}

