% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_ballot}
\alias{clean_ballot}
\title{Master one-step cleaning function}
\usage{
clean_ballot(ballot, lookup, format)
}
\arguments{
\item{ballot}{The raw ballot image data frame}

\item{lookup}{The raw lookup image data frame}

\item{format}{A character string detailing the format. Current
supported formats are "WinEDS" and "ChoicePlus" (in progress), based on
common types of software used. Contact creators with suggestions for
more formats.}
}
\value{
The ballot data, but human-readable
}
\description{
Wraps `label` and `characterize` to clean the ballot
image in one step.
}
\examples{
clean_ballot(ballot = sf_bos_ballot, lookup = sf_bos_lookup, format = "WinEDS")
}
