% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{clean_mn}
\alias{clean_mn}
\title{Function for cleaning Minneapolis RCV data}
\usage{
clean_mn(data)
}
\arguments{
\item{data}{The raw RCV data}
}
\value{
The data formatted for use with rcv_tally
}
\description{
The Minneapolis data comes in a different form than SF or Cambridge data.
This function optimizes the process for formatting this data.
}
\examples{
clean_mn(minneapolis_mayor_2013)
}
