% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{elected}
\alias{elected}
\title{Summarizes which candidates were elected from the results tally}
\usage{
elected(results, n = 1)
}
\arguments{
\item{results}{the results tabulated from the election in question}

\item{n}{the number of candidates being elected (defaults to 1)}
}
\value{
a vector of the candidates successfully elected
}
\description{
Summarizes which candidates were elected from the results tally
}
