% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{label}
\alias{label}
\title{Separates single column election data frames.}
\usage{
label(data, image, format)
}
\arguments{
\item{data}{A data frame with a single column}

\item{image}{Whether the data is a "ballot" or "lookup" image}

\item{format}{A character string detailing the format. Current
supported formats are "WinEDS" (used in San Francisco and Alameda counties)
and "ChoicePlus" (in progress, used in Cambridge, MA), based on common types
of software used. Contact the creators with suggestions for more formats.}
}
\value{
A data frame with multiple columns
}
\description{
Takes a data frame of a single column (i.e. sf_bos_ballot) and splits it
into usable named columns.
}
\examples{
label(data = sf_bos_ballot, image = "ballot", format = "WinEDS")
}
