% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{label_ballot}
\alias{label_ballot}
\title{Separates single column ballot data frames.}
\usage{
label_ballot(data, method)
}
\arguments{
\item{data}{A data frame containing the ballot image data}

\item{method}{A character string detailing the method. Current
supported methods are "WinEDS" and "ChoicePlus" (forthcoming), based on
common types of software used. Contact creators with suggestions for
more methods.}
}
\value{
A data frame with multiple columns
}
\description{
Takes a data frame of a single column (i.e. sf_bos_ballot) and splits it
into usable named columns.
}
\examples{
label_ballot(data = sf_bos_ballot)
}
