% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{make_alluvialdf}
\alias{make_alluvialdf}
\title{Creates a data frame for use with the alluvial package}
\usage{
make_alluvialdf(image, rcvcontest)
}
\arguments{
\item{image}{A dataframe containing rcv election data}

\item{rcvcontest}{The election to calculate results for}
}
\value{
A dataframe that counts how many ballots follow each unique "path"
of candidates through the election rounds
}
\description{
Creates a data frame for use with the alluvial package
}
\examples{
make_alluvialdf(image = sf_bos_clean, rcvcontest = "Board of Supervisors, District 1")
}
