% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{make_d3list}
\alias{make_d3list}
\title{Creates a data frame for use with the networkD3 package}
\usage{
make_d3list(results)
}
\arguments{
\item{results}{A dataframe containing rcv election results in the format
produced by `rcv_tally()`}
}
\value{
A list of 2 dataframes that can be used to construct sankey diagrams
via the networkD3 package's `sankeyNetwork()` function
}
\description{
Creates a data frame for use with the networkD3 package
}
\examples{
make_d3list(results = sf_7_results)
}
