% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{minneapolis_mayor_2013}
\alias{minneapolis_mayor_2013}
\title{Raw ballot data from a Minneapolis, MN RCV election}
\format{A data frame with 80101 rows and 5 variables}
\source{
\url{http://vote.minneapolismn.gov/www/groups/public/@clerk/documents/webcontent/2013-mayor-cvr.xlsx}
}
\usage{
minneapolis_mayor_2013
}
\description{
The .rda version of a .csv from a 2013 Minneapolis, MN mayoral RCV
election. This data is included as an example of another ballot format
for use with ballot functions.
}
\keyword{datasets}

