% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{old_ballot_tidy}
\alias{old_ballot_tidy}
\title{Old function for reading in raw CA rcv ballot images}
\usage{
old_ballot_tidy(rawimage, masterkey)
}
\arguments{
\item{rawimage}{A .csv or similar file containing the raw ballot image}

\item{masterkey}{A .csv or similar file containing the lookup file}
}
\value{
A dataframe that can be easily read and used for calculating results
}
\description{
Old function for reading in raw CA rcv ballot images
}
\examples{
old_ballot_tidy(rawimage = "http://www.acgov.org/rov/rcv/results/230/BerkeleyMayor/ballot_image.txt",
masterkey = "http://www.acgov.org/rov/rcv/results/230/BerkeleyMayor/master_lookup.txt")
}
