% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{old_rcv_plot}
\alias{old_rcv_plot}
\title{Old function for creating a single style of sankey plot from rcv results}
\usage{
old_rcv_plot(rcvresults)
}
\arguments{
\item{rcvresults}{A specifically formatted dataframe of rcv results}
}
\value{
A sankey plot
}
\description{
Old function for creating a single style of sankey plot from rcv results
}
\examples{
old_rcv_plot(mayorresults)
}

