% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{old_rcv_tally}
\alias{old_rcv_tally}
\title{Old function for determining rcv election results}
\usage{
old_rcv_tally(rcvimage, rcvcontest)
}
\arguments{
\item{rcvimage}{A specifically formatted dataframe of an rcv ballot image}

\item{rcvcontest}{The specific contest that one wants to determine results for}
}
\value{
A dataframe containing round by round sums
}
\description{
Old function for determining rcv election results
}
\examples{
old_rcv_tally(rcvimage = sf_bos_clean, rcvcontest = "Board of Supervisors, District 1")
}
