\name{rcv-package}
\alias{rcv-package}
\alias{rcv}
\docType{package}
\title{
\packageTitle{rcv}
}
\description{
\packageDescription{rcv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rcv}
\packageIndices{rcv}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rcv}

Maintainer: \packageMaintainer{rcv}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
