% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv-description.R
\docType{package}
\name{rcv}
\alias{rcv}
\alias{rcv-package}
\title{rcv: A package for determining ranked choice election results.}
\description{
Work in progress package with a few work in progress functions.
}
\section{rcv Functions}{

import_data(), label(), characterize(), clean_ballot(),
readable(), rcv_tally()
}

