% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulation.R
\name{rcv_tally}
\alias{rcv_tally}
\title{Determines RCV round results in a dataframe}
\usage{
rcv_tally(image, rcvcontest)
}
\arguments{
\item{image}{A dataframe containing rcv election data}

\item{rcvcontest}{(optional) The election to calculate results for. If the image
contains more than one unique contest, this must be supplied. In most cases
except for some San Francisco elections, this is unnecessary.}
}
\value{
A dataframe that contains vote tallies
}
\description{
Determines RCV round results in a dataframe
}
\examples{
rcv_tally(image = sf_bos_clean, rcvcontest = "Board of Supervisors, District 7")
}
