% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_functions.R
\name{rcv_tally_test}
\alias{rcv_tally_test}
\title{First attempt at getting RCV results with a prebuilt dataframe}
\usage{
rcv_tally_test(image, rcvcontest)
}
\arguments{
\item{image}{A dataframe containing rcv election data}

\item{rcvcontest}{The election to calculate results for}
}
\value{
A dataframe that contains vote tallies
}
\description{
First attempt at getting RCV results with a prebuilt dataframe
}
\examples{
rcv_tally_test(image = "sf_bos_clean", rcvcontest = "Board of Supervisors, District 1")
}
