% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{readable}
\alias{readable}
\title{Alters ballot dataframe to have each row correspond to a single voter}
\usage{
readable(clean)
}
\arguments{
\item{clean}{A tidy dataframe that comes from ballot_tidy()}
}
\value{
A dataframe that can be easily read and understood by humans
}
\description{
Alters ballot dataframe to have each row correspond to a single voter
}
\examples{
readable(sf_bos_clean)
}
