% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{sf_7_results}
\alias{sf_7_results}
\title{Tabulated results from a San Francisco RCV election}
\format{a data frame with 6 rows and 5 variables}
\usage{
sf_7_results
}
\description{
Results from the SF "Board of Supervisors, District 7" election
in 2016. Tabulated with rcv_tally.
rcv_tally(image = sf_bos_clean,
          rcvcontest = "Board of Supervisors, District 7")
}
\keyword{datasets}
