% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{sf_ballot_labelled}
\alias{sf_ballot_labelled}
\title{Labelled ballot data from San Francisco RCV election}
\format{A data frame with 643806 rows and 9 variables}
\usage{
sf_ballot_labelled
}
\description{
The .rda version of sf_bos_ballot.rda after the `label()`
function has been applied to it.
}
\keyword{datasets}
