% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{sf_bos_ballot}
\alias{sf_bos_ballot}
\title{Raw ballot image data from a San Francisco RCV election}
\format{A data frame with 643806 rows and 1 variable}
\source{
\url{http://www.sfelections.org/results/20161108/data/20161206/20161206_ballotimage.txt}
}
\usage{
sf_bos_ballot
}
\description{
The .rda version of a raw .txt file, containing ballot data
in numeric form for the 2016 San Francisco Board of Supervisors
elections in Districts 1, 3, 5, 7, 9, and 11.
}
\keyword{datasets}
