% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{sf_bos_lookup}
\alias{sf_bos_lookup}
\title{Raw ballot lookup data from a San Francisco RCV election.}
\format{A data frame with 644 rows and 1 variable}
\source{
\url{http://www.sfelections.org/results/20161108/data/20161206/20161206_masterlookup.txt}
}
\usage{
sf_bos_lookup
}
\description{
The .rda version of a raw .txt file, containing numeric and
character information to match the numeric data in the raw
ballot image with the objects these numeric strings refer to.
Data refers to the 2016 San Francisco Board of Supervisors
elections in Districts 1, 3, 5, 7, 9, and 11.
}
\keyword{datasets}
