% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{sf_lookup_labelled}
\alias{sf_lookup_labelled}
\title{Labelled ballot lookup data from a San Francisco RCV election.}
\format{A data frame with 644 rows and 7 variables}
\usage{
sf_lookup_labelled
}
\description{
The .rda version of sf_bos_lookup.rda after the `label()`
function has been applied to it.
}
\keyword{datasets}
