% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_functions.R
\name{label_ballot}
\alias{label_ballot}
\title{Separates single column ballot data frames.}
\usage{
label_ballot(data)
}
\arguments{
\item{data}{A data frame containing the ballot image data}
}
\value{
A data frame with multiple columns
}
\description{
Takes a data frame of a single column (i.e. sf_bos_ballot) and splits it
into usable named columns.
}
\examples{
label_ballot(data = sf_bos_ballot)
}
