% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv-description.R
\docType{package}
\name{rcv}
\alias{rcv}
\alias{rcv-package}
\title{Package: rcv}
\description{
rcv: A package for determining ranked choice election results.
}
\details{
Work in progress package with a few work in progress functions.
}
\section{rcv Functions}{

import_data(), label(), characterize(), clean_ballot(), clean_mn(),
readable(), elected(), approval(), rcv_tally(), add_exhausted(),
make_alluvialdf(), make_d3list()
}

