% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv-description.R
\docType{package}
\name{rcv}
\alias{rcv}
\title{Package: rcv}
\description{
Ranked Choice Voting
}
\details{
A collection of ranked choice voting data and functions to
manipulate, run elections with, and visualize this data and others.
It can bring in raw data, transform it into a ballot you can read,
and return election results for an RCV contest.
}
\section{rcv Functions}{

label(), characterize(), clean_ballot(), clean_mn(),
readable(), elected(), approval(), rcv_tally(), add_exhausted(),
make_alluvialdf(), make_d3list()
}

